windows/requirements_windows.txt
flask>=3.0.0
requests>=2.31.0
tzdata>=2024.1
pystray>=0.19.5
Pillow>=10.0.0
